color_avg_computer/requirements.txt
asyncio-nats-client==0.11.4
opencv-python==4.5.2.52
aiounittest==1.4.0
